third_party/blink/web_tests/external/wpt/css/css-syntax/input-preprocessing-expected.txt

This is a testharness.js-based test.
[FAIL] "fooU+d800" becomes "foo�"
  assert_equals: expected "foo\\ufffd" but got "fooU+d800"
[FAIL] "fU+d800oo" becomes "f�oo"
  assert_equals: expected "f\\ufffdoo" but got "fU+d800oo"
[FAIL] "U+d800foo" becomes "�foo"
  assert_equals: expected "\\ufffdfoo" but got "U+d800foo"
[FAIL] "U+d800" becomes "�"
  assert_equals: expected "\\ufffd" but got "U+d800"
[FAIL] "U+d800U+d800U+d800" becomes "���"
  assert_equals: expected "\\ufffd\\ufffd\\ufffd" but got "U+d800U+d800U+d800"
Harness: the test ran to completion.
